FEATURE: Support nested and standalone Psmb.Ajaxify:RenderPath
A key feature of this package is to associate a partial with a unique
partial key for later dedicated rendering in AJAX requests. The
implementation strictly relied on a defined position of the responsible
Psmb.Ajaxify:RenderPath within the Psmb.Ajaxify:Ajaxify Fusion
object. With this patch, that relationship has been relaxed: the
Psmb.Ajaxify:RenderPath Fusion object can now be used both standalone
and nested within any depth of Psmb.Ajaxify:Ajaxify. However, when
used standalone, it must be called at the first level of the partial -
in the same way as Psmb.Ajaxify:Ajaxify.