[!!!]FEATURE: Auto-generate unique key of partial
The partial key must be unique per Fusion path and content node.
Therefore, the current state of one key per Fusion path is not
sufficient. With this patch, the key will be a hash based on both
attributes.

The hash is generated automatically, thereby removing the requirement
for the developer to manually specify a unique key.

Because this patch changes the value of the publicly used URL parameter
`&ajaxPathKey=<value>` this patch is marked as breaking.