[!!!]TASK: Refactor
Previously, partial rendering considered only the Fusion path. With the
recent changes, both the Fusion path and the content node, called the
partial context, are now considered. Consider this change in the naming
conventions. In detail these are:

- Public: Rename URL parameter "ajaxPathKey" to "ajaxPartialKey".
- Public: Rename `Psmb.Ajaxify:RenderPath` to `Psmb.Ajaxify:PartialSerializer`.
- Public: Rename `Psmb.Ajaxify:RenderPathResolver` to `Psmb.Ajaxify:PartialResolver`.
- Public: Rename "[P]athsCache" to "[Pp]artialCache" cache identifier.
- Private: Use FQDN for \Exception to satisfy the editor :).
- Private: Use verbose common "fusionPath" over "path" and "renderPath".
- Private: Rename "pathKey" to "partialKey".
- Private: Even smaller adjustments ..

Since this patch changes the naming of some public elements, it is marked
as breaking. And breaking changes are included because there were other
breaking changes in this patch already.